pyproject.toml
[project]
name = "pokemon-scouting"
version = "0.1.0"
description = ""
authors = [
    {name = "alelles16@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
flake8 = "^7.2.0"
black = "^25.1.0"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
  \.git
 | \.venv
 | build
 | dist
)/
'''
